services/llm-service/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-json-logger>=2.0.7
pyyaml>=6.0
httpx>=0.26.0
